% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{calculate_relevance_multiclass}
\alias{calculate_relevance_multiclass}
\title{Calculate Relevance for Multiclass Classification}
\usage{
calculate_relevance_multiclass(
  X,
  y,
  fdr_level = 0.05,
  hypotheses_independent = FALSE,
  correction_method = "fdr",
  n_significant = 1,
  test_binary_real = "mann"
)
}
\arguments{
\item{X}{Feature matrix}

\item{y}{Target vector}

\item{fdr_level}{FDR threshold}

\item{hypotheses_independent}{Independence assumption}

\item{correction_method}{Multiple testing correction method}

\item{n_significant}{Minimum classes to be significant for}

\item{test_binary_real}{Test for binary target + real feature}
}
\value{
data.frame with multiclass relevance information
}
\description{
Calculate Relevance for Multiclass Classification
}
\keyword{internal}
