% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{classify_feature}
\alias{classify_feature}
\title{Classify Feature Type}
\usage{
classify_feature(feature)
}
\arguments{
\item{feature}{A numeric vector}
}
\value{
"constant", "binary", or "real"
}
\description{
Classify Feature Type
}
\keyword{internal}
