% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_total_correlation.R
\name{compute_joint_entropy}
\alias{compute_joint_entropy}
\title{Compute Joint Entropy of Multivariate Time Series}
\usage{
compute_joint_entropy(X, bins = 10)
}
\arguments{
\item{X}{Numeric matrix (N x T)}

\item{bins}{Integer, number of bins for discretization}
}
\value{
Joint entropy value
}
\description{
Compute Joint Entropy of Multivariate Time Series
}
\keyword{internal}
