% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_synchronization.R
\name{compute_mutual_information}
\alias{compute_mutual_information}
\title{Compute mutual information between two vectors}
\usage{
compute_mutual_information(x, y, bins = 10)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{bins}{Number of bins for histogram (default 10)}
}
\value{
Mutual information estimate
}
\description{
Simple histogram-based estimator of mutual information
}
\keyword{internal}
