% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_diversity.R
\name{compute_shannon_entropy_simple}
\alias{compute_shannon_entropy_simple}
\title{Compute Shannon entropy from binned time series}
\usage{
compute_shannon_entropy_simple(x, bins = 10)
}
\arguments{
\item{x}{Numeric vector}

\item{bins}{Number of bins for histogram (default 10)}
}
\value{
Shannon entropy estimate
}
\description{
Simple histogram-based Shannon entropy estimator
}
\keyword{internal}
