% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cross_spectral_density}
\alias{cpp_cross_spectral_density}
\title{Compute cross-spectral density for a pair of series - Rcpp version}
\usage{
cpp_cross_spectral_density(x, y)
}
\arguments{
\item{x}{First time series}

\item{y}{Second time series}
}
\value{
List with coherence, phase, and CSD magnitude
}
\description{
Uses FFT to compute cross-spectral density
}
\keyword{internal}
