% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_crossing_points}
\alias{cpp_crossing_points}
\title{Number of crossing points (median crossings)}
\usage{
cpp_crossing_points(x)
}
\arguments{
\item{x}{Numeric vector (time series)}
}
\value{
Number of median crossings
}
\description{
Computes the number of times a time series crosses its median
}
