% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_embed2_incircle}
\alias{cpp_embed2_incircle}
\title{2D embedding - proportion of points inside circle}
\usage{
cpp_embed2_incircle(y, boundary)
}
\arguments{
\item{y}{Numeric vector (time series)}

\item{boundary}{Circular boundary (1 or 2)}
}
\value{
Proportion of points inside circle
}
\description{
Points inside given circular boundary in 2-d embedding space
}
