% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_flat_spots}
\alias{cpp_flat_spots}
\title{Longest flat spot}
\usage{
cpp_flat_spots(x)
}
\arguments{
\item{x}{Numeric vector (time series)}
}
\value{
Maximum run length in any bin
}
\description{
Maximum run length within any of 10 equal-sized bins
}
