% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_motiftwo_entro3}
\alias{cpp_motiftwo_entro3}
\title{Binary motif entropy (3-letter words)}
\usage{
cpp_motiftwo_entro3(y)
}
\arguments{
\item{y}{Numeric vector (time series)}
}
\value{
Entropy of binary motifs
}
\description{
Entropy of 3-letter binary words (above/below mean)
}
