% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mutual_information}
\alias{cpp_mutual_information}
\title{Compute mutual information between two vectors - Rcpp version}
\usage{
cpp_mutual_information(x, y, bins = 10L)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{bins}{Number of bins for histogram}
}
\value{
Mutual information estimate
}
\description{
Histogram-based estimator of mutual information
}
\keyword{internal}
