% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mv_shannon_entropy}
\alias{cpp_mv_shannon_entropy}
\title{Compute Shannon entropy from binned data - multivariate version}
\usage{
cpp_mv_shannon_entropy(x, bins = 10L)
}
\arguments{
\item{x}{Numeric vector}

\item{bins}{Number of bins}
}
\value{
Shannon entropy
}
\description{
Compute Shannon entropy from binned data - multivariate version
}
\keyword{internal}
