% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_pairwise_ccf}
\alias{cpp_pairwise_ccf}
\title{Compute pairwise cross-correlation function (CCF) features - Rcpp version}
\usage{
cpp_pairwise_ccf(X, max_lag)
}
\arguments{
\item{X}{Matrix (N x T) where rows are series, columns are timepoints}

\item{max_lag}{Maximum lag for CCF}
}
\value{
List with ccf_max_vals, ccf_lag_vals, ccf_zero_vals for all pairs
}
\description{
Efficiently computes CCF for all pairs of time series
}
\keyword{internal}
