% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_pairwise_mi}
\alias{cpp_pairwise_mi}
\title{Compute pairwise mutual information - Rcpp version}
\usage{
cpp_pairwise_mi(X, bins = 10L)
}
\arguments{
\item{X}{Matrix (N x T) where rows are series, columns are timepoints}

\item{bins}{Number of bins for histogram}
}
\value{
Vector of mutual information values for all pairs
}
\description{
Efficiently computes MI for all pairs of time series
}
\keyword{internal}
