% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_spectral_shape}
\alias{cpp_spectral_shape}
\title{Compute spectral shape descriptors}
\usage{
cpp_spectral_shape(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with spectral shape features
}
\description{
Compute spectral shape descriptors
}
