% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_walker_propcross}
\alias{cpp_walker_propcross}
\title{Walker crossing proportion}
\usage{
cpp_walker_propcross(y)
}
\arguments{
\item{y}{Numeric vector (time series)}
}
\value{
Proportion of crossings between walker and time series
}
\description{
Simulates a walker moving through time domain
}
