% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_sets_registry.R
\name{get_feature_sets}
\alias{get_feature_sets}
\title{Get Available Feature Sets}
\usage{
get_feature_sets()
}
\value{
Character vector of feature set names
}
\description{
Returns a character vector of all available feature set names.
}
\examples{
# List all available feature sets
feature_sets <- get_feature_sets()
head(feature_sets)
}
