% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_utils.R
\name{get_multivariate_feature_sets}
\alias{get_multivariate_feature_sets}
\title{Get names of multivariate feature sets}
\usage{
get_multivariate_feature_sets()
}
\value{
Character vector of available feature set names
}
\description{
Get names of multivariate feature sets
}
\keyword{internal}
