% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{infer_ml_task}
\alias{infer_ml_task}
\title{Infer Machine Learning Task}
\usage{
infer_ml_task(y)
}
\arguments{
\item{y}{Target vector}
}
\value{
"classification" or "regression"
}
\description{
Infer Machine Learning Task
}
\keyword{internal}
