% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_utils.R
\name{standardize_multivariate}
\alias{standardize_multivariate}
\title{Standardize multivariate time series}
\usage{
standardize_multivariate(X)
}
\arguments{
\item{X}{Matrix (N x T)}
}
\value{
Standardized matrix
}
\description{
Standardizes each series (row) to have mean 0 and sd 1
}
\keyword{internal}
