% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{test_binary_binary}
\alias{test_binary_binary}
\title{Test Binary Target + Binary Feature}
\usage{
test_binary_binary(feature, target)
}
\arguments{
\item{feature}{Binary feature vector}

\item{target}{Binary target vector}
}
\value{
p-value
}
\description{
Uses Fisher's exact test
}
\keyword{internal}
