% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{test_real_binary}
\alias{test_real_binary}
\title{Test Real Target + Binary Feature}
\usage{
test_real_binary(feature, target)
}
\arguments{
\item{feature}{Binary feature vector}

\item{target}{Real-valued target vector}
}
\value{
p-value
}
\description{
Uses Kolmogorov-Smirnov test
}
\keyword{internal}
