% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{test_real_real}
\alias{test_real_real}
\title{Test Real Target + Real Feature}
\usage{
test_real_real(feature, target)
}
\arguments{
\item{feature}{Real-valued feature vector}

\item{target}{Real-valued target vector}
}
\value{
p-value
}
\description{
Uses Kendall's tau correlation test
}
\keyword{internal}
