% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_acf}
\alias{ts_acf}
\title{Extract autocorrelation features}
\usage{
ts_acf(x, max_lag = 20)
}
\arguments{
\item{x}{Numeric vector}

\item{max_lag}{Maximum lag for ACF calculation}
}
\value{
Named list of ACF features
}
\description{
Extract autocorrelation features
}
