% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_acf_summary}
\alias{ts_acf_summary}
\title{ACF summary statistics}
\usage{
ts_acf_summary(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with acf_mean_abs_1to10, acf_decay_index
}
\description{
Mean absolute ACF and decay index.
}
