% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_autocov_zero_lag_norm}
\alias{ts_autocov_zero_lag_norm}
\title{Normalized zero-lag autocovariance}
\usage{
ts_autocov_zero_lag_norm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with acov0_norm
}
\description{
Ratio of variance to mean squared.
}
