% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_basic}
\alias{ts_basic}
\title{Extract basic statistical features}
\usage{
ts_basic(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list of basic statistics
}
\description{
Computes basic statistics: maximum, minimum, and root mean square.
}
