% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_bds}
\alias{ts_bds}
\title{BDS test for nonlinearity}
\usage{
ts_bds(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with bds_stat, bds_pvalue
}
\description{
Correlation integral-based test for independence.
}
