% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_block_var_ratio}
\alias{ts_block_var_ratio}
\title{Block variance ratio}
\usage{
ts_block_var_ratio(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with var_ratio_blocks
}
\description{
Ratio of max to min variance across blocks.
}
