% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_const_segments}
\alias{ts_const_segments}
\title{Constant segments}
\usage{
ts_const_segments(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with const_seg_count, const_seg_max_len
}
\description{
Count and max length of constant segments.
}
