% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_crossings_mean_median}
\alias{ts_crossings_mean_median}
\title{Crossings around mean and median}
\usage{
ts_crossings_mean_median(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with crossings_mean, crossings_median
}
\description{
Counts of crossings around mean and median.
}
