% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_cusum_sq}
\alias{ts_cusum_sq}
\title{CUSUM of squared deviations}
\usage{
ts_cusum_sq(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with cusum_sq_range
}
\description{
Range of cumulative sum of squared deviations.
}
