% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_diff}
\alias{ts_diff}
\title{Extract differencing features}
\usage{
ts_diff(x, include_ndiffs = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{include_ndiffs}{Logical, whether to include ndiffs (requires urca package)}
}
\value{
Named list of differencing ACF/PACF features
}
\description{
Extract differencing features
}
