% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_diff_stats}
\alias{ts_diff_stats}
\title{Difference statistics}
\usage{
ts_diff_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with mean_abs_diff, var_diff, kurt_diff
}
\description{
Descriptors on first differences.
}
