% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_dwell_times}
\alias{ts_dwell_times}
\title{Dwell times}
\usage{
ts_dwell_times(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with dwell_above_mean, dwell_below_mean
}
\description{
Mean run lengths above and below median.
}
