% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_entropy}
\alias{ts_entropy}
\title{Extract entropy features}
\usage{
ts_entropy(x, m = 2, r = 0.2)
}
\arguments{
\item{x}{Numeric vector}

\item{m}{Embedding dimension for sample/approximate entropy}

\item{r}{Tolerance for sample/approximate entropy}
}
\value{
Named list of entropy features
}
\description{
Extract entropy features
}
