% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_entropy_dispersion}
\alias{ts_entropy_dispersion}
\title{Dispersion entropy}
\usage{
ts_entropy_dispersion(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with dispersion_entropy
}
\description{
Robust entropy using value-to-class mapping.
}
