% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_entropy_rate_order1}
\alias{ts_entropy_rate_order1}
\title{First-order entropy rate}
\usage{
ts_entropy_rate_order1(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with entropy_rate1
}
\description{
Conditional entropy H(X|X-1) from transition matrix.
}
