% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_event_rate_over_threshold}
\alias{ts_event_rate_over_threshold}
\title{Event rate over threshold}
\usage{
ts_event_rate_over_threshold(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with event_rate_t, mean_exceedance_t
}
\description{
Rate and mean exceedance over threshold.
}
