% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_extrema}
\alias{ts_extrema}
\title{Extract extrema location features}
\usage{
ts_extrema(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list of extrema locations
}
\description{
Computes normalized locations (0 to 1) of first and last occurrences
of maximum and minimum values.
}
