% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_fft}
\alias{ts_fft}
\title{Extract FFT coefficient features}
\usage{
ts_fft(x, num_coef = 3)
}
\arguments{
\item{x}{Numeric vector}

\item{num_coef}{Number of FFT coefficients to extract (default 3)}
}
\value{
Named list of FFT features
}
\description{
Extract FFT coefficient features
}
