% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_forecastability}
\alias{ts_forecastability}
\title{Forecastability measure}
\usage{
ts_forecastability(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with forecastability
}
\description{
Computes 1 - normalized spectral entropy.
}
