% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_fourier1}
\alias{ts_fourier1}
\title{Fourier harmonic features}
\usage{
ts_fourier1(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with harmonic_amp1, harmonic_phase1
}
\description{
Amplitude and phase of first harmonic.
}
