% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_fractal_higuchi}
\alias{ts_fractal_higuchi}
\title{Higuchi fractal dimension}
\usage{
ts_fractal_higuchi(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with higuchi_fd
}
\description{
Fractal dimension via average curve length scaling.
}
