% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_fractal_katz}
\alias{ts_fractal_katz}
\title{Katz fractal dimension}
\usage{
ts_fractal_katz(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with katz_fd
}
\description{
Measures complexity via polyline distance metric.
}
