% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_fractal_petrosian}
\alias{ts_fractal_petrosian}
\title{Petrosian fractal dimension}
\usage{
ts_fractal_petrosian(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with petrosian_fd
}
\description{
Fractal dimension based on sign changes in differences.
}
