% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_gini}
\alias{ts_gini}
\title{Gini coefficient}
\usage{
ts_gini(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with gini_coefficient
}
\description{
Measure of statistical dispersion.
}
