% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_goodness_normality}
\alias{ts_goodness_normality}
\title{Goodness-of-fit to normality}
\usage{
ts_goodness_normality(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with ad_stat, cvm_stat
}
\description{
Anderson-Darling and Cramér-von Mises statistics.
}
