% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_gph_d}
\alias{ts_gph_d}
\title{Geweke-Porter-Hudak long-memory estimate}
\usage{
ts_gph_d(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with gph_d
}
\description{
Estimates long-range dependence parameter via periodogram regression.
}
