% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_harmonicity_index}
\alias{ts_harmonicity_index}
\title{Harmonicity index}
\usage{
ts_harmonicity_index(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with harmonicity_top3
}
\description{
Ratio of top 3 peaks to total power.
}
