% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_hjorth}
\alias{ts_hjorth}
\title{Hjorth parameters}
\usage{
ts_hjorth(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with activity, mobility, complexity
}
\description{
Computes Hjorth activity, mobility, and complexity parameters.
}
