% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_iat}
\alias{ts_iat}
\title{Integrated autocorrelation time}
\usage{
ts_iat(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with int_autocorr_time
}
\description{
Sum of ACF until first nonpositive or threshold.
}
