% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_iqr_props}
\alias{ts_iqr_props}
\title{IQR-based properties}
\usage{
ts_iqr_props(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with iqr_over_median_abs, outlier_prop_iqr
}
\description{
IQR relative to median and outlier proportion.
}
