% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_line_length}
\alias{ts_line_length}
\title{Line length}
\usage{
ts_line_length(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with line_length
}
\description{
Sum of absolute differences (total variation).
}
