% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_mase}
\alias{ts_mase}
\title{Mean Absolute Scaled Error}
\usage{
ts_mase(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with mase, smase
}
\description{
In-sample forecast error normalized by naive forecast MAE.
}
