% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_misc}
\alias{ts_misc}
\title{Extract miscellaneous features}
\usage{
ts_misc(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list of miscellaneous features
}
\description{
Computes additional useful features including ACF zero crossing,
zero proportion, and first derivative statistics.
}
