% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_monotonicity}
\alias{ts_monotonicity}
\title{Monotonicity}
\usage{
ts_monotonicity(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with kendall_tau_abs, mk_pvalue
}
\description{
Kendall correlation with time index and Mann-Kendall p-value.
}
