% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_multipeak_psd}
\alias{ts_multipeak_psd}
\title{Multiple peaks in PSD}
\usage{
ts_multipeak_psd(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with psd_num_peaks, psd_peak_contrast
}
\description{
Count and contrast of spectral peaks.
}
