% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_multiscale_perm}
\alias{ts_multiscale_perm}
\title{Multiscale permutation entropy}
\usage{
ts_multiscale_perm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with mpe_2, mpe_3, mpe_5
}
\description{
Ordinal-pattern entropy over coarse-grained scales.
}
