% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_correlation.R
\name{ts_mv_correlation}
\alias{ts_mv_correlation}
\title{Correlation structure features for multivariate time series}
\usage{
ts_mv_correlation(X, method = "pearson", ...)
}
\arguments{
\item{X}{Matrix (N x T) where N is number of series, T is timepoints}

\item{method}{Character; correlation method. One of "pearson" (default),
"spearman", or "kendall"}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Named list of 15 correlation-based features
}
\description{
Extracts features characterizing the correlation structure between time series.
}
\examples{
\dontrun{
# 4 time series with varying correlations
X <- matrix(rnorm(4 * 150), nrow = 4)
X[2,] <- 0.8 * X[1,] + 0.2 * rnorm(150)
corr_features <- ts_mv_correlation(X)
}
}
