% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_covariance.R
\name{ts_mv_covariance}
\alias{ts_mv_covariance}
\title{Covariance matrix features for multivariate time series}
\usage{
ts_mv_covariance(X)
}
\arguments{
\item{X}{Matrix (N x T) where N is number of series, T is timepoints}
}
\value{
Named list of 7 covariance-based features
}
\description{
Extracts features based on the covariance matrix of the time series.
Unlike correlation features which use standardized data, these use the
actual variances and covariances.
}
\examples{
\dontrun{
# 3 time series with different variances
X <- matrix(0, nrow = 3, ncol = 150)
X[1,] <- rnorm(150, sd = 1)
X[2,] <- rnorm(150, sd = 2)
X[3,] <- rnorm(150, sd = 0.5)
cov_features <- ts_mv_covariance(X)
}
}
