% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_pca.R
\name{ts_mv_pca}
\alias{ts_mv_pca}
\title{PCA-based multivariate features}
\usage{
ts_mv_pca(X)
}
\arguments{
\item{X}{Matrix (N x T) where N is number of series, T is timepoints}
}
\value{
Named list of 15 PCA-based features
}
\description{
Extracts features based on principal component analysis of the time series.
Performs PCA on the correlation matrix of the N time series.
}
\examples{
\dontrun{
# 5 correlated time series
X <- matrix(rnorm(5 * 200), nrow = 5)
X[2,] <- 0.7 * X[1,] + 0.3 * rnorm(200)
pca_features <- ts_mv_pca(X)
}
}
