% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_pacf}
\alias{ts_pacf}
\title{Extract PACF features}
\usage{
ts_pacf(x, max_lag = 20)
}
\arguments{
\item{x}{Numeric vector}

\item{max_lag}{Maximum lag for PACF calculation}
}
\value{
Named list of PACF features
}
\description{
Extract PACF features
}
