% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_page_hinkley}
\alias{ts_page_hinkley}
\title{Page-Hinkley test}
\usage{
ts_page_hinkley(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with ph_stat, ph_cp_index
}
\description{
Change point detection statistic and index.
}
