% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_peak_symmetry}
\alias{ts_peak_symmetry}
\title{Peak symmetry statistics}
\usage{
ts_peak_symmetry(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with peak_symmetry_mean, peak_symmetry_sd
}
\description{
Mean and SD of peak symmetry (left vs right area).
}
