% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_pettitt}
\alias{ts_pettitt}
\title{Pettitt change point test}
\usage{
ts_pettitt(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with pettitt_stat, pettitt_pval, pettitt_cp_index
}
\description{
Non-parametric test for change point in distribution.
}
