% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_qcd}
\alias{ts_qcd}
\title{Quartile coefficient of dispersion}
\usage{
ts_qcd(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with quartile_coef_dispersion
}
\description{
Normalized measure of spread based on quartiles.
}
