% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_quantile_spread_trend}
\alias{ts_quantile_spread_trend}
\title{Quantile spread trend}
\usage{
ts_quantile_spread_trend(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with qsp_slope
}
\description{
Slope of rolling IQR over time.
}
