% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_renyi2_entropy}
\alias{ts_renyi2_entropy}
\title{Rényi entropy (order 2)}
\usage{
ts_renyi2_entropy(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with renyi2
}
\description{
Rényi entropy of order 2 from histogram.
}
