% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_return_stats}
\alias{ts_return_stats}
\title{Return statistics}
\usage{
ts_return_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with mean_abs_ret, vol_ret, skew_ret
}
\description{
Descriptors on returns (relative changes).
}
