% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_robust_shape}
\alias{ts_robust_shape}
\title{Robust shape statistics}
\usage{
ts_robust_shape(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with bowley_skew, moors_kurtosis
}
\description{
Bowley skewness and Moors kurtosis based on quantiles.
}
