% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_runs}
\alias{ts_runs}
\title{Run-length and up-down statistics}
\usage{
ts_runs(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with run statistics
}
\description{
Computes statistics based on runs above/below median.
}
