% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_shape_factors}
\alias{ts_shape_factors}
\title{Wave and impulse shape factors}
\usage{
ts_shape_factors(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with shape factors
}
\description{
Computes crest, impulse, shape, and clearance factors.
}
