% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_spectral_crest}
\alias{ts_spectral_crest}
\title{Spectral crest factor}
\usage{
ts_spectral_crest(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with spectral_crest_factor
}
\description{
Ratio of peak to mean spectral power.
}
