% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_spectral_psr}
\alias{ts_spectral_psr}
\title{Peak-to-sidelobe ratio}
\usage{
ts_spectral_psr(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with peak_sidelobe_ratio
}
\description{
Ratio of dominant peak to second largest peak in spectrum.
}
