% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_spectral_qfactor}
\alias{ts_spectral_qfactor}
\title{Spectral Q-factor}
\usage{
ts_spectral_qfactor(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with spectral_q
}
\description{
Quality factor of dominant spectral peak.
}
