% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_spectral_slope_r2}
\alias{ts_spectral_slope_r2}
\title{Spectral slope R²}
\usage{
ts_spectral_slope_r2(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with spectral_slope_r2
}
\description{
R² of log-log spectral slope fit.
}
