% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_ssc}
\alias{ts_ssc}
\title{Slope sign changes}
\usage{
ts_ssc(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with slope_sign_changes
}
\description{
Number of slope sign changes in the series.
}
