% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_stats}
\alias{ts_stats}
\title{Extract statistical features}
\usage{
ts_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list of statistical features
}
\description{
Extract statistical features
}
