% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_structure}
\alias{ts_structure}
\title{Extract structure features}
\usage{
ts_structure(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list of structure features
}
\description{
Extract structure features
}
