% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_tail_asymmetry}
\alias{ts_tail_asymmetry}
\title{Tail asymmetry}
\usage{
ts_tail_asymmetry(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with upper_tail_ratio, lower_tail_ratio
}
\description{
Upper and lower tail ratios.
}
