% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_tail_index}
\alias{ts_tail_index}
\title{Hill tail index}
\usage{
ts_tail_index(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with hill_tail_index
}
\description{
Estimates tail heaviness via Hill estimator.
}
