% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_tail_ratios}
\alias{ts_tail_ratios}
\title{Tail ratio statistics}
\usage{
ts_tail_ratios(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with q99_q95, q95_q90
}
\description{
Ratios of extreme quantiles.
}
