% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_tkao}
\alias{ts_tkao}
\title{Teager-Kaiser energy operator}
\usage{
ts_tkao(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with teager_mean, teager_std
}
\description{
Mean and std of Teager-Kaiser energy.
}
