% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_trimmed_stats}
\alias{ts_trimmed_stats}
\title{Trimmed statistics}
\usage{
ts_trimmed_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with trimmed_mean_p05, trimmed_sd_p05
}
\description{
Trimmed mean and SD at 5% level.
}
