% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_turning_points}
\alias{ts_turning_points}
\title{Turning points}
\usage{
ts_turning_points(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with turning_point_rate
}
\description{
Rate of local extrema in the series.
}
