% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_value_range}
\alias{ts_value_range}
\title{Value range over standard deviation}
\usage{
ts_value_range(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with range_over_std
}
\description{
Range normalized by standard deviation.
}
