% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_volatility_cluster}
\alias{ts_volatility_cluster}
\title{Volatility clustering}
\usage{
ts_volatility_cluster(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with ac_abs_l1, ac_sq_l1, ac_sq_l5, ac_sq_l10
}
\description{
ACF of absolute and squared values.
}
