% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_winsor_stats}
\alias{ts_winsor_stats}
\title{Winsorized statistics}
\usage{
ts_winsor_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with winsor_mean, winsor_var
}
\description{
Mean and variance after Winsorizing at 5%.
}
