% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_zcr_diff}
\alias{ts_zcr_diff}
\title{Zero crossing rate of differences}
\usage{
ts_zcr_diff(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named list with zcr_diff
}
\description{
Rate of sign changes in first differences.
}
