% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_utils.R
\name{validate_multivariate_input}
\alias{validate_multivariate_input}
\title{Validate and convert multivariate input to matrix format}
\usage{
validate_multivariate_input(X)
}
\arguments{
\item{X}{Either a matrix (N x T) or list of N numeric vectors}
}
\value{
Matrix with N rows (series) and T columns (timepoints)
}
\description{
Validate and convert multivariate input to matrix format
}
\keyword{internal}
